More tidying up.
