+++ src/cli/did-verification-key-revoke.ts
@@ -4,7 +4,7 @@ import { writeFile } from 'node:fs/promises';
 import { parseArgs } from 'node:util';
 import { PlcClientError } from '@did-plc/lib';
 import { importRotationKeyPair } from '../keys.js';
-import { revokeVerificationKey } from '../did.js';
+import { revokeVerificationKey } from '../plc.js';
 import { loadRotationKey, SigningKeyError, KeyData } from '../signing.js';
 import { logPlcError } from './lib/plc-error.js';
 import { rotationKeyHelp } from './lib/help.js';
